// src/store/mutation-types.ts
export enum MutationTypes {
    ADD_TO_CART = 'addToCart',
    DELETE_FROM_CART = 'deleteFromCart',
    ADD_PRODUCT = 'addProduct',
    DELETE_PRODUCT = 'deleteProduct',
    SET_PRODUCTS = 'setProducts',
    SET_VAT_RATES = 'setVatRates',
    SET_VAT_RATE = 'setVatRate',
    SET_EXCHANGE_RATE = 'setExchangeRate',
    SET_CURRENCY = 'setCurrency',
}